+++ matlab/changelog.md
@@ -3,6 +3,14 @@ IronClust, written by J. James Jun, Flatiron Institute, Simons Foundation
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [4.2.4] - 2018-11-02
+### Added
+- `irc sbatch-mda` uses Flatiron compute-cluster (sbatch + disBatch)
+  - syntax: `irc batch-mda input_dir output_dir (template_file)`
+- `irc batch-mda` runs 
+  - syntax: `irc batch-mda input_dir output_dir (template_file)`
+
+
 ## [4.2.3] - 2018-11-02
 ### Added
 - Dockerfile added